Adds additional examples on implemented interfaces
To continue the long line of examples. I wanted to try two
more cases that I didn't consider before. First analyzing a
the invocation of the method of a class that implemented a non
generic interface and also passing such a class to a method
expecting an interface of the implemented type. The second
example was the same but with a generic interface. The problem
persists in the case of generic interfaces passed to methods.
NullPointerException is the error shown.