Adds additional examples using lambds in Spark
Recently I found a way on how to analyze lambdas in SPF. It was
necessary to point to the lambda${0seq} private (and normally
static) method created inside a class. SPF behaves as expected.

The new example is completely written using lambdas, which resembles
more to the real way spark programs are written.